refactor: :poop: Remove some pyqt releated code in core_reef.py
